feat(mob) : création des mobs avec position random

+++ code/Drones/Model/mob.cs
@@ -19,8 +19,8 @@ namespace BigNightmare
             _y = y;
         }
 
-        public float X { get { return _x; } set { _x = value; } }
-        public float Y { get { return _y; } set { _y = value; } }
+        public int X { get { return (int)_x; } set { _x = value; } }
+        public int Y { get { return (int)_y; } set { _y = value; } }
 
         public void Update(int interval)
         {
